move session secret helpers to runtimeconfig
