+++ runtimeconfig/session_secret_test.go
@@ -1,17 +1,15 @@
-package core_test
+package runtimeconfig
 
 import (
 	"path/filepath"
 	"testing"
 
 	"github.com/arran4/goa4web/config"
-	"github.com/arran4/goa4web/core"
 	common "github.com/arran4/goa4web/core/common"
-	"github.com/arran4/goa4web/runtimeconfig"
 )
 
 func TestLoadSessionSecretCLI(t *testing.T) {
-	secret, err := core.LoadSessionSecret(core.OSFS{}, "cli", "", config.EnvSessionSecret, config.EnvSessionSecretFile)
+	secret, err := LoadSessionSecret(useMemFS(t), "cli", "", config.EnvSessionSecret, config.EnvSessionSecretFile)
 	if err != nil {
 		t.Fatalf("unexpected error: %v", err)
 	}
@@ -22,7 +20,7 @@ func TestLoadSessionSecretCLI(t *testing.T) {
 
 func TestLoadSessionSecretEnv(t *testing.T) {
 	t.Setenv(config.EnvSessionSecret, "env")
-	secret, err := core.LoadSessionSecret(core.OSFS{}, "", "", config.EnvSessionSecret, config.EnvSessionSecretFile)
+	secret, err := LoadSessionSecret(useMemFS(t), "", "", config.EnvSessionSecret, config.EnvSessionSecretFile)
 	if err != nil {
 		t.Fatalf("unexpected error: %v", err)
 	}
@@ -32,12 +30,12 @@ func TestLoadSessionSecretEnv(t *testing.T) {
 }
 
 func TestLoadSessionSecretFile(t *testing.T) {
-	fs := core.UseMemFS(t)
+	fs := useMemFS(t)
 	file := "sec"
 	if err := fs.WriteFile(file, []byte("fromfile"), 0600); err != nil {
 		t.Fatalf("write file: %v", err)
 	}
-	secret, err := core.LoadSessionSecret(fs, "", file, config.EnvSessionSecret, config.EnvSessionSecretFile)
+	secret, err := LoadSessionSecret(fs, "", file, config.EnvSessionSecret, config.EnvSessionSecretFile)
 	if err != nil {
 		t.Fatalf("unexpected error: %v", err)
 	}
@@ -47,9 +45,9 @@ func TestLoadSessionSecretFile(t *testing.T) {
 }
 
 func TestLoadSessionSecretGenerate(t *testing.T) {
-	fs := core.UseMemFS(t)
+	fs := useMemFS(t)
 	file := "new"
-	secret, err := core.LoadSessionSecret(fs, "", file, config.EnvSessionSecret, config.EnvSessionSecretFile)
+	secret, err := LoadSessionSecret(fs, "", file, config.EnvSessionSecret, config.EnvSessionSecretFile)
 	if err != nil {
 		t.Fatalf("unexpected error: %v", err)
 	}
@@ -68,7 +66,7 @@ func TestLoadSessionSecretGenerate(t *testing.T) {
 func TestDefaultSessionSecretPathDev(t *testing.T) {
 	common.Version = "dev"
 	t.Setenv(config.EnvDocker, "")
-	got := runtimeconfig.DefaultSessionSecretPath()
+	got := DefaultSessionSecretPath()
 	if got != ".session_secret" {
 		t.Fatalf("want .session_secret got %s", got)
 	}
@@ -78,7 +76,7 @@ func TestDefaultSessionSecretPathDocker(t *testing.T) {
 	common.Version = "1"
 	t.Setenv(config.EnvDocker, "1")
 	t.Setenv("HOME", "/home/test")
-	got := runtimeconfig.DefaultSessionSecretPath()
+	got := DefaultSessionSecretPath()
 	if got != "/var/lib/goa4web/session_secret" {
 		t.Fatalf("want /var/lib/goa4web/session_secret got %s", got)
 	}
@@ -88,7 +86,7 @@ func TestDefaultSessionSecretPathUser(t *testing.T) {
 	common.Version = "1"
 	t.Setenv(config.EnvDocker, "")
 	t.Setenv("XDG_CONFIG_HOME", "/cfg")
-	got := runtimeconfig.DefaultSessionSecretPath()
+	got := DefaultSessionSecretPath()
 	want := filepath.Join("/cfg", "goa4web", "session_secret")
 	if got != want {
 		t.Fatalf("want %s got %s", want, got)
